Added socket.io functionality, client-side logic for legal moves.
Users connected to the website can now send each other moves, given that
the start from the same initial configuration.
Added images and basic html and css styling to display the board and
moves next to it.
Added client-side logic to determine which moves are legal. Currently,
this includes only king and pawn moves, which are the easiest and
hardest to implement respectively.