Can now match black and white in queue and play.
Implemented 'match' signals that will be sent to each player, assigning
them their colour and prompting a new game to start.

There is a bug where the black player only starts their game when a file
is updated so that nodemon starts again. Also when four players join,
thy are matched in pairs but moves still get sent to everyone.