Users are now correctly paired.
If four users join, then we get two independent games.
Bug persists where black's pieces (the even player in the queue) are not
shown at the start of the game. Probably not loaded in time for the
initial drawing of the board. Clicking on a square where a black piece
should be loads the board correctly.