Fixed castling checkmate bug. Added mate to movelist.
The issue was that the ChessGame object handled castling by doing two
separate moves: first the king then the rook. This caused the move-list
to increment twice and the active color to switch twice.
As a result, the opponent thought they had "no legal moves" left because they thought it was not their turn and all moves appeared illegal.