Fixed pawn move of Issue #5.
Issue was that the pawn pattern function in isLegal.js checked for
"enPasssantPosition exists OR move moves into enPassantPosition" rather
than "AND".
Also, the enpassantposition itself was always passed into the function
as a string such as "a6" rather than a matrix position such as [0,5].
This was also fixed to enable correct enpassant behaviour.